fix(exports): fail the render when the export produced no content
render_png_export documents that bytes are None only alongside asset.exception,
but a workflow reporting success while storing nothing fell through with neither,
leaving callers unable to tell that apart from a successful render. Record it as
an exception so the contract holds, and treat empty content the same as missing
content when falling back to object storage, matching get_content_response.

Also moves TestInsightQueryWantsLegend next to the function it covers in
image_exporter, rather than in the facade tests where it was left when the legend
decision moved out of the facade.

Generated-By: PostHog Code
Task-Id: 9778184f-3ca5-4170-a6f9-ac8b6fd01b3e
